Moved logic from login page to App.tsx
