+++ src/App.tsx
@@ -3,7 +3,8 @@ import {
   Resource,
   CustomRoutes,
   Loading,
-  type DataProvider
+  type DataProvider,
+  type AuthProvider
 } from 'react-admin'
 import { Route } from 'react-router-dom'
 import MyLayout from './components/Layout'
@@ -37,10 +38,33 @@ function App(): React.ReactElement {
   const [dataProvider, setDataProvider] = useState<DataProvider | undefined>(
     undefined
   )
+  const [authProvider, setAuthProvider] = useState<AuthProvider | undefined>(
+    undefined
+  )
   const [loggingPref, setLoggingPref] = useState<boolean>(false)
   const handleGetProvider = (): any => {
-    if (loggingPref !== null)
-      getDataProvider(loggingPref).then(setDataProvider).catch(console.log)
+    if (loggingPref !== null) {
+      getDataProvider(loggingPref)
+        .then((provider) => {
+          setDataProvider(provider)
+          const authenticationProvider = autProvider(provider)
+          setAuthProvider(authenticationProvider)
+          if (provider !== undefined && dataProvider === undefined) {
+            const queryParams = new URLSearchParams(window.location.search)
+            const username = queryParams.get('username')
+            const password = queryParams.get('password')
+            if (username !== null && password !== null) {
+              authenticationProvider
+                .login({ username, password })
+                .then((_) => {
+                  window.history.replaceState({}, '', window.location.pathname)
+                })
+                .catch(console.log)
+            }
+          }
+        })
+        .catch(console.log)
+    }
   }
 
   useEffect(() => {
@@ -65,6 +89,7 @@ function App(): React.ReactElement {
 
   useEffect(handleGetProvider, [loggingPref])
   if (dataProvider === undefined) return LoadingPage
+  if (authProvider === undefined) return LoadingPage
 
   return (
     <Suspense fallback={LoadingPage}>
@@ -72,7 +97,7 @@ function App(): React.ReactElement {
         dashboard={Welcome}
         dataProvider={dataProvider}
         loginPage={Login}
-        authProvider={autProvider(dataProvider)}
+        authProvider={authProvider}
         layout={MyLayout}
         theme={rcoTheme}
         disableTelemetry
