+++ src/pages/Login.tsx
@@ -15,7 +15,6 @@ import {
   TableRow
 } from '@mui/material'
 import AppIcon from '../assets/rco_transparent.png'
-import { useEffect } from 'react'
 
 export default function Login() {
   const login = useLogin()
@@ -30,21 +29,6 @@ export default function Login() {
       notify('Invalid email or password', { type: 'error' })
     })
   }
-
-  useEffect(() => {
-    const queryParams = new URLSearchParams(window.location.search)
-    const username = queryParams.get('username')
-    const password = queryParams.get('password')
-    if (username !== null && password !== null) {
-      login({ username, password })
-        .then((_) => {
-          window.history.replaceState({}, '', window.location.pathname)
-        })
-        .catch(() => {
-          notify('Invalid email or password', { type: 'error' })
-        })
-    }
-  }, [])
 
   return (
     <>
